Added tint and blending mode customization per icon

+++ core/common/src/main/java/org/elnix/dragonlauncher/common/utils/ImageUtils.kt
@@ -321,7 +321,7 @@ object ImageUtils {
         // Step 4: color tint
         icon.tint?.let {
             paint.colorFilter = PorterDuffColorFilter(
-                it.toInt(),
+                it,
                 PorterDuff.Mode.SRC_IN
             )
         }
